+++ Q3/Dicke.m
@@ -1,23 +1,24 @@
 (* -*- mode:math -*- *)
 (* Mahn-Soo Choi *)
-(* $Date: 2021-01-14 11:41:29+09 $ *)
-(* $Revision: 1.6 $ *)
+(* $Date: 2021-01-28 11:24:46+09 $ *)
+(* $Revision: 1.9 $ *)
 
 
 BeginPackage[ "Q3`Dicke`",
   { "Q3`Fock`", "Q3`Quisso`",
-    "Q3`Pauli`", "Q3`Cauchy`", "Q3`Abel`" }
+    "Q3`Pauli`", "Q3`Cauchy`", "Q3`" }
  ]
 
 Unprotect[Evaluate[$Context<>"*"]]
 
-Print @ StringJoin[
+Begin["`Private`"]
+`Version = StringJoin[
   $Input, " v",
-  StringSplit["$Revision: 1.6 $"][[2]], " (",
-  StringSplit["$Date: 2021-01-14 11:41:29+09 $"][[2]], ") ",
+  StringSplit["$Revision: 1.9 $"][[2]], " (",
+  StringSplit["$Date: 2021-01-28 11:24:46+09 $"][[2]], ") ",
   "Mahn-Soo Choi"
- ]
-
+ ];
+End[]
 
 { DickeBasis };
 
@@ -47,8 +48,7 @@ Protect[ Evaluate @ $symbs ]
 End[]
 
 
-Q3`Dicke`Private`symbs = Protect[Evaluate[$Context<>"*"]]
+Q3Protect[]
 
-SetAttributes[Evaluate[Q3`Dicke`Private`symbs], ReadProtected]
 
 EndPackage[]