+++ Q3/Gray.m
@@ -4,25 +4,31 @@
    based on the Gray code. *)
 
 (* Mahn-Soo Choi *)
-(* $Date: 2021-01-17 12:09:22+09 $ *)
-(* $Revision: 1.16 $ *)
+(* $Date: 2021-02-14 00:58:47+09 $ *)
+(* $Revision: 1.26 $ *)
 
 BeginPackage[ "Q3`Gray`",
-  { "Q3`Quisso`", "Q3`Pauli`", "Q3`Cauchy`", "Q3`Abel`" }
+  { "Q3`Quisso`", "Q3`Pauli`", "Q3`Cauchy`", "Q3`" }
  ]
 
 Unprotect[Evaluate[$Context<>"*"]]
 
-Print @ StringJoin[
+Begin["`Private`"]
+`Version = StringJoin[
   $Input, " v",
-  StringSplit["$Revision: 1.16 $"][[2]], " (",
-  StringSplit["$Date: 2021-01-17 12:09:22+09 $"][[2]], ") ",
+  StringSplit["$Revision: 1.26 $"][[2]], " (",
+  StringSplit["$Date: 2021-02-14 00:58:47+09 $"][[2]], ") ",
   "Mahn-Soo Choi"
- ]
+ ];
+End[]
+
+{ BinaryToGray, GrayToBinary };
 
 { GrayCode, GrayCodeSubsets };
 
 { GrayControlledU, GrayControlledW };
+
+{ GrayTwoLevelU, TwoLevelU, TwoLevelDecomposition };
 
 
 Begin["`Private`"]
@@ -30,17 +36,7 @@ Begin["`Private`"]
 $symbs = Unprotect[]
 
 
-GrayCode::usage = "GrayCode[n] returns the n-bit Gray code in binary digits."
-
-GrayCode[1] = {{0}, {1}}
-
-GrayCode[n_Integer] := Join[
-  Map[Prepend[#, 0] &, GrayCode[n - 1]],
-  Map[Prepend[#, 1] &, Reverse@GrayCode[n - 1]]
- ] /; n > 1
-
-
-GrayCodeSubsets::usage = "GrayCodeSubsets[l] constructs a binary reflected Gray code on set l."
+GrayCodeSubsets::usage = "GrayCodeSubsets[set] constructs a binary-reflected Gray code on set, that is, returns the list of all subsets of set each of which differs from its predecessor by only one element."
 
 (* NOTE: The code has just been copied from Combinatorica package. *)
 
@@ -52,7 +48,7 @@ GrayCodeSubsets[l_List] := Block[
   { $RecursionLimit = Infinity,
     s }, 
   s = GrayCodeSubsets[Take[l, 1-Length[l]]];
-  Join[s,  Map[Prepend[#, First[l]] &, Reverse[s]]]
+  Join[s,  Map[Prepend[#, First[l]]&, Reverse[s]]]
  ]
 
 
@@ -133,13 +129,150 @@ grayCtrl[{aa__?QubitQ, b_?QubitQ}, op_, lbl_] := With[
  ]
 
 
+BinaryToGray::usage = "BinaryToGray[bits] converts the binary number bits to a bit-reflected Gray code (BRGC)."
+
+(* https://en.wikipedia.org/wiki/Gray_code *)
+BinaryToGray[bits_List] := Mod[bits + ShiftRight[bits], 2]
+
+GrayToBinary::usage = "GrayToBinary[gray] converts the bit-reflected Gray code gray to a binary number."
+
+(* https://en.wikipedia.org/wiki/Gray_code *)
+GrayToBinary[gg : {0 ..}] := gg
+
+(* https://en.wikipedia.org/wiki/Gray_code *)
+GrayToBinary[gray_List] := Module[
+  { j = First @ FirstPosition[gray, 1],
+    mask },
+  mask = Total @ Table[ShiftRight[gray, n], {n, 1, Length[gray] - j}];
+  Mod[gray + mask, 2]
+ ]
+
+
+GrayCode::usage = "GrayCode[n, x_, y_] returns the list of n-bit Gray codes connecting the integers x and y.\nEach Gray code is given in binary digits.\nGrayCode[n] returns the list of all n-bit Gray codes.\nGrayCode[n] is equivalent to GrayCode[n, 0, 2^n-1]."
+
+GrayCode[n_Integer, x_, y_] := Module[
+  { a = FromDigits[GrayToBinary[IntegerDigits[x, 2, n]], 2],
+    b = FromDigits[GrayToBinary[IntegerDigits[y, 2, n]], 2] },
+  If[ a > b,
+    Reverse[ BinaryToGray /@ IntegerDigits[Range[b, a], 2, n] ],
+    BinaryToGray /@ IntegerDigits[Range[a, b], 2, n]
+   ]
+ ]
+
+GrayCode[1] = {{0}, {1}}
+
+GrayCode[n_Integer] := Join[
+  Map[Prepend[#, 0]&, GrayCode[n - 1]],
+  Map[Prepend[#, 1]&, Reverse @ GrayCode[n - 1]]
+ ] /; n > 1
+
+
+GrayTwoLevelU::usage = "GrayTwoLevelU[mat, {x, y}, {q1, q2, ...}] returns a list of CNOT gates and one controlled-U gate that compose the two-level unitary operation specified by mat."
+
+GrayTwoLevelU[ TwoLevelU[mat_?MatrixQ, {x_, y_}, L_], qq:{__?QubitQ} ] :=
+  GrayTwoLevelU[mat, {x, y}, FlavorNone @ qq] /;
+  L == Power[2, Length @ qq]
+
+GrayTwoLevelU[mat_?MatrixQ, {x_Integer, y_Integer}, qq:{__?QubitQ}] :=
+  Module[
+    { gray = GrayCode[Length @ qq, x-1, y-1],
+      mask, expr },
+    mask = Catenate @ Successive[grayCNOT[#1, #2, qq]&, Most @ gray];
+    expr = grayCtrlU[gray[[-2]], gray[[-1]], mat, qq];
+    Join[mask, expr, Reverse @ mask]
+   ] /; OrderedQ @ {x, y}
+
+grayCNOT[g1_, g2_, qq_] := Module[
+  { xx = g1 + g2,
+    cc = Mod[g1 + g2, 2] },
+  xx = Pick[qq, xx, 0];
+  xx = Through[xx[1]];
+  cc = GroupBy[Transpose @ {cc, qq}, First, Last @* Transpose];
+  {xx, CNOT[cc[0], cc[1]], xx} /. {} -> Nothing
+ ]
+
+grayCtrlU[g1_, g2_, mat_, qq_] := Module[
+  { xx = g1 + g2,
+    cc = Mod[g1 + g2, 2],
+    op },
+  xx = Pick[qq, xx, 0];
+  xx = Through[xx[1]];
+  cc = GroupBy[Transpose @ {cc, qq}, First, Last @* Transpose];
+  op = QuissoExpression[mat, cc[1]];
+  If[ Not @ OrderedQ @ {g1, g2},
+    With[
+      { tt = First @ cc[1] },
+      op = Multiply[tt[1], op, tt[1]]
+     ]
+   ];
+  {xx, ControlledU[cc[0], op, "Label"->"U"], xx} /. {} -> Nothing
+ ]
+
+
+TwoLevelU::usage = "TwoLevelU[mat, {i, j}, n] represents a two-level unitary matrix, that is, the 2 x 2 unitary matrix mat operating on the ith and jth rows and columns of an n x n matrix.\nMatrix[TwoLevelU[mat, {i, j}, n]] returns the full n x n matrix.\nSee also GrayTwoLevelU, which decomposes TwoLevelU[...] into CNOT gates and one controlled-U gate."
+
+TwoLevelU /:
+HoldPattern @ Matrix @ TwoLevelU[mat_?MatrixQ, {x_, y_}, n_] := Module[
+  { new = One[n],
+    val = Flatten @ mat,
+    idx = Tuples[{x, y}, 2] },
+  ReplacePart[new, Thread[idx -> val]]
+ ]
+
+
+TwoLevelDecomposition::usage = "TwoLevelDecomposition[mat] returns a list of two-level unitary matrices U1, U2, ... in terms of TwoLevelU, where Dot[U1, U2, ...] is formally equivalent to mat."
+
+TwoLevelDecomposition[mat_?MatrixQ] := twoLevelDCMP[mat, 1]
+
+twoLevelDCMP[mat_?MatrixQ, j_Integer] := Module[
+  { mm = Rest @ mat,
+    UU, vv },
+  {vv, UU} = twoLevelDCMP[First @ mat, j];
+  If[ j == Length[First @ mat], Return @ UU ];
+  If[ mm == {}, Return @ UU ];
+  mm = Dot[ mm, Sequence @@ Reverse[ Topple /@ Matrix /@ UU ] ];
+  Join[ twoLevelDCMP[mm, j+1], UU ]
+ ]
+
+twoLevelDCMP[vec_?VectorQ, k_Integer] := Module[
+  { new, UU, U },
+  {new, UU} = twoLevelDCMP[vec, k+1];
+  {new, U}  = twoLevelDCMP[new, {k}];
+  {new, Prepend[UU, U]}
+ ] /; 1 <= k < Length[vec]-1
+
+twoLevelDCMP[vec_?VectorQ, k_Integer] := Module[
+  {new, U},
+  {new, U} = twoLevelDCMP[vec, {k}];
+  {new, {U}}
+ ] /; k >= Length[vec]-1
+
+twoLevelDCMP[vec_?VectorQ, {k_Integer}] := Module[
+  { new = Take[vec, {k, k+1}],
+    L = Length[vec],
+    U },
+  U = {
+    new,
+    {-1, 1} Reverse[Conjugate @ new]
+   } / Norm[new];
+  
+  new = ReplacePart[vec, {k -> Norm[new], k+1 -> 0}];
+  { new, TwoLevelU[U, {k, k+1}, L] }
+ ] /; 1 <= k < Length[vec]
+
+twoLevelDCMP[vec_?VectorQ, {k_Integer}] := With[
+  { z = Last @ vec },
+  { ReplacePart[vec, k -> Abs @ z],
+    TwoLevelU[DiagonalMatrix @ {1, z/Abs[z]}, {k-1, k}, k] }
+ ] /; k == Length[vec]
+
+
 Protect[ Evaluate @ $symbs ]
 
 End[]
 
 
-Q3`Gray`Private`symbs = Protect[Evaluate[$Context<>"*"]]
+Q3Protect[]
 
-SetAttributes[Evaluate[Q3`Gray`Private`symbs], ReadProtected]
 
 EndPackage[]