+++ Q3/PacletInfo.m
@@ -1,13 +1,14 @@
 (* Paclet Info File *)
 
-(* created 2021/01/19*)
+(* created 2021/02/11*)
 
 Paclet[
     Name -> "Q3",
-    Version -> "1.4.2",
+    Version -> "1.5.0",
     MathematicaVersion -> "12+",
     Description -> "Mathematica application to help study quantum information processing, quantum many-body systems, and quantum spin systems.",
     Creator -> "Mahn-Soo Choi (Korea University)",
+    WolframVersion -> "12+",
     Extensions -> 
         {
             {"Documentation", Language -> "English"}